Updated client project structure, moved and reformatted some files
