+++ client/src/views/App.tsx
@@ -1,0 +1,24 @@
+import React, { FC, ReactElement } from 'react';
+import logo from '../assets/images/logo.svg';
+import '../assets/styles/App.css';
+
+const App: FC = (): ReactElement => (
+  <div className="App">
+    <header className="App-header">
+      <img src={logo} className="App-logo" alt="logo" />
+      <p>
+        Edit <code>src/App.tsx</code> and save to reload.
+      </p>
+      <a
+        className="App-link"
+        href="https://reactjs.org"
+        target="_blank"
+        rel="noopener noreferrer"
+      >
+        Learn React
+      </a>
+    </header>
+  </div>
+);
+
+export default App;
